aside/mini-projects/radar/pyproject.toml
[tool.poetry]
name = "airforce"
version = "0.1.0"
description = ""
authors = ["selimslab <hselimozturk@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.26.0"
python-dotenv = "^0.19.1"
pandas = "^1.3.4"
matplotlib = "^3.4.3"
seaborn = "^0.11.2"
mplcursors = "^0.5"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
isort = "^5.10.1"
black = "^21.10b0"
mypy = "^0.910"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
